Fixed problems with default entries in the message test utility
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@5959 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-jini/src/java/com/bigdata/service/jini/util/MessageTest.java
@@ -180,13 +180,13 @@ public class MessageTest {
         final int parallel = (Integer) fed
                 .getClient()
                 .getConfiguration()
-                .getEntry(COMPONENT, ConfigurationOptions.PARALLEL, Long.class,
+                .getEntry(COMPONENT, ConfigurationOptions.PARALLEL, Integer.class,
                         ConfigurationOptions.DEFAULT_PARALLEL);
         
         final int payload = (Integer) fed
                 .getClient()
                 .getConfiguration()
-                .getEntry(COMPONENT, ConfigurationOptions.PAYLOAD, Long.class,
+                .getEntry(COMPONENT, ConfigurationOptions.PAYLOAD, Integer.class,
                         ConfigurationOptions.DEFAULT_PAYLOAD);
 
         try {
